+++ packages/adhere/changelog/CHANGELOG.md
@@ -5,6 +5,8 @@
 2022-6-5
 
 * 按需加载antd
+* **adhere-ui-conditionalrender**
+  - 加入conditionalNotEmptyArr方法
 * **adhere-ui-messagedialog**
   - Modal参数名称修改(cloneBtn修改为closeBtn)
 * **adhere-ui-scrollload**